requirements.txt
cloudinary==1.1.1
cssmin==0.2.0

django-admin-tools==0.5.2
django-appconf==1.0.1
django-ckeditor==4.4.7
django-compressor==1.4
django-contrib-comments==1.5
django-filebrowser==3.5.7
django-grappelli==2.5.6
django-htmlmin==0.7.0
django-mptt==0.6.1
django-pipeline==1.5.3
django-redis-cache==0.13.1
django-redis-sessions==0.4.0
django-social-share==0.3.0
django-staticfiles==1.2.1
django-tagging==0.3.4
django-wpadmin==1.7.0
django-xmlrpc==0.1.5
futures==3.0.3
lxml==3.4.3
html5lib==0.999
lxml==3.4.3
oauthlib==0.7.2
Pillow==2.7.0
ply==3.4
pyparsing==2.0.3
python-memcached==1.54
python-openid==2.2.5
pytz==2014.10
redis==2.10.3
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
slimit==0.8.1
wheel==0.24.0
